Refactoring the elasticSearchUtil one implementation class
