+++ metacat-main/src/main/java/com/netflix/metacat/main/services/ServicesModule.java
@@ -21,9 +21,9 @@ import com.netflix.metacat.main.services.impl.MViewServiceImpl;
 import com.netflix.metacat.main.services.impl.PartitionServiceImpl;
 import com.netflix.metacat.main.services.impl.TableServiceImpl;
 import com.netflix.metacat.main.services.search.ElasticSearchUtil;
+import com.netflix.metacat.main.services.search.ElasticSearchUtilImpl;
 import com.netflix.metacat.main.services.search.ElasticSearchClientProvider;
 import com.netflix.metacat.main.services.search.MetacatEventHandlers;
-import com.netflix.metacat.main.services.search.ElasticSearchUtilProvider;
 import com.netflix.metacat.main.services.search.ElasticSearchMetacatRefresh;
 import org.elasticsearch.client.Client;
 
@@ -44,7 +44,7 @@ public class ServicesModule extends AbstractModule {
         //search
         bind(Client.class).toProvider(ElasticSearchClientProvider.class).in(Singleton.class);
         binder().bind(MetacatEventHandlers.class).in(Singleton.class);
-        bind(ElasticSearchUtil.class).toProvider(ElasticSearchUtilProvider.class).in(Singleton.class);
+        binder().bind(ElasticSearchUtil.class).to(ElasticSearchUtilImpl.class).in(Singleton.class);
         binder().bind(ElasticSearchMetacatRefresh.class).in(Singleton.class);
         binder().bind(MetadataService.class).in(Singleton.class);
     }
